requirements.txt
# Dependencias del proyecto API de Productos
# Django y Django REST Framework
Django==5.2.6
djangorestframework==3.16.1

# Base de datos MySQL
mysqlclient==2.2.7

# Variables de entorno
python-dotenv==1.1.1

# Documentación API
drf-spectacular==0.28.0

# Dependencias adicionales de Django
asgiref==3.9.2
sqlparse==0.5.3
tzdata==2025.2

# Dependencias de drf-spectacular
uritemplate==4.2.0
PyYAML==6.0.2
jsonschema==4.25.1
inflection==0.5.1
attrs==25.3.0
jsonschema-specifications==2025.9.1
referencing==0.36.2
rpds-py==0.27.1
typing-extensions==4.15.0
